#99 - Initial setup, statistics filtered by subject and generalisation of verify token middleware for reusability.
